+++ HISTORY.md
@@ -1,6 +1,7 @@
 # Release History
 
 ## Current snapshot
+* Add setJobStatus and setLastUpdate to JobRepository interface and both implementations - This fixes a race condition during keepalive and message
 
 ## Release 0.74.0
 * Removed dependency to edison-hal (was introduced for testing purposes only).